+++ ChangeLog.md
@@ -4,6 +4,7 @@ Change Log for the Setup cookbook
 Version 1.1.11
 -------------
  - Avoid issue where installing NetManager will brick networking (#29)
+ - Add support for installing openstack clients from the Fedora RDO repo.
 
 Version 1.1.10
 -------------